ath6kl: Add large IE in CONNECT event
Firmware API version 4 supports an upgrade of the size of IEs that can
be transmitted in a CONNECT event response from 255 bytes to 65535
bytes. However, the new size is NOT opt-in, meaning that the previous
driver that didn't recognize the longer 16-bit length fields would
improperly process the IEs and cause things like WPA handshakes to fail.

This patch adds support for the large IE capability while maintaining
backwards compatability with the "normal" size IEs.

https://jira.digi.com/browse/ALLEG-42

(cherry picked from commit edf09905ceefa5e3fa3155c6ba21e8b45bc2ff13)
(cherry picked from commit 9466ac6bff10318c241ee50cf7006766481ea654)
(cherry picked from commit aea3cb931ba0921ec9b520753f93ef9ba842a357)

+++ drivers/net/wireless/ath/ath6kl/cfg80211.c
@@ -736,8 +736,8 @@ void ath6kl_cfg80211_connect_event(struct ath6kl_vif *vif, u16 channel,
 				   u8 *bssid, u16 listen_intvl,
 				   u16 beacon_intvl,
 				   enum network_type nw_type,
-				   u8 beacon_ie_len, u8 assoc_req_len,
-				   u8 assoc_resp_len, u8 *assoc_info)
+				   u16 beacon_ie_len, u16 assoc_req_len,
+				   u16 assoc_resp_len, u8 *assoc_info)
 {
 	struct ieee80211_channel *chan;
 	struct ath6kl *ar = vif->ar;
